requirements.txt
bidict 
more-itertools 
numpy 
pandas 
simple-parsing
structlog 
tensorflow
typing-extensions 
jsonconversion
python-dotenv